requirements.txt
discord.py
python-dotenv==1.0.1
google-generativeai==0.8.3
Flask==3.0.3
